+++ app/main.js
@@ -2,6 +2,7 @@ const { app, BrowserWindow, Tray, ipcMain, Notification } = require('electron')
 const path = require('path')
 const open = require('open')
 const appMenu = require('./app-menu')
+const request = require('./node/request')
 
 const isDev = process.env.NODE_ENV === 'development'
 
@@ -101,4 +102,19 @@ ipcMain.on('exit', () => {
 
 ipcMain.on('openURL', (e, url) => {
   open(url)
+})
+
+ipcMain.on('request', (e, requestParams) => {
+  request(requestParams).then((data) => {
+    if (data.status === 401) {
+      e.sender.send('request-401')
+      return
+    }
+
+    return data.json()
+  }).then(json => e.sender.send('request-success', json))
+    .catch((err) => {
+      console.log(err)
+      e.sender && e.sender.send('request-failed')
+    })
 })
