+++ package.json
@@ -29,6 +29,7 @@
     "electron-store": "^5.1.1",
     "iscroll": "^5.2.0",
     "jquery": "^3.4.1",
+    "node-fetch": "^2.6.0",
     "open": "^7.0.3"
   }
 }
